Use $this->setMwGlobals() to set test globals / configs
Instead of utilizing the HashConfig class, it's better to
make use of the $this->setMwGlobals() method to set globals
used in test cases. This is provided by MediaWikiTestCase
and also their previous configs are restored after test runs
and parent::tearDown() is called.

Also, removed some unused namespaces per the refactoring of
the setConfig() method.

Change-Id: Ida2addd23aa94c7f71c9f1eee98f91a8b838f7e8